Moving ListTagsRequest into Query folder.
